requirements.txt
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.8
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.1.0
Flask-SocketIO==5.4.1
fonttools==4.55.0
graphviz==0.20.3
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
pyarrow==18.1.0
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
threadpoolctl==3.5.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
websockets>=12.0
Werkzeug==3.1.3
wsproto==1.2.0